feat: add socialify badge and update description
Add a socialify badge to the README for better visual appeal and update
the
description to reflect the purpose of the scripts more accurately.
This change
introduces a new way to engage with the project's community
through Telegram
and provides a glimpse into the benefits of using the
scripts for dailyroutines.

BREAKING CHANGE: None.