docs(docusaurus.config.js): change sidebar label to '快速开始'
Update the sidebar label from '文档' to '快速开始' to improve the user
experience for Chinese users. This change makes the documentation more
accessible by using a more common and intuitive label.

docs/support/index.md: add donation support page

Create a new page to encourage users to donate to support the project. The
page includes information on why donations are important and how they will
be used. It also features a section for manual record-keeping of donors and
amounts. This addition aims to increase engagement and sustainability of
the project.

BREAKING CHANGE: None.